requirements.txt
asgiref==3.4.1
Django==3.2.5
django-crum==0.7.9
pytz==2021.1
sqlparse==0.4.1
